+++ viser/__init__.py
@@ -1,4 +1,3 @@
-from ._gui_api import Gui3dContainerHandle as Gui3dContainerHandle
 from ._gui_api import GuiFolderHandle as GuiFolderHandle
 from ._gui_api import GuiTabGroupHandle as GuiTabGroupHandle
 from ._gui_api import GuiTabHandle as GuiTabHandle
@@ -9,6 +8,7 @@ from ._gui_handles import GuiHandle as GuiHandle
 from ._icons_enum import Icon as Icon
 from ._scene_handles import CameraFrustumHandle as CameraFrustumHandle
 from ._scene_handles import FrameHandle as FrameHandle
+from ._scene_handles import Gui3dContainerHandle as Gui3dContainerHandle
 from ._scene_handles import ImageHandle as ImageHandle
 from ._scene_handles import LabelHandle as LabelHandle
 from ._scene_handles import MeshHandle as MeshHandle
