+++ viser/client/src/ControlPanel/Generated.tsx
@@ -31,14 +31,10 @@ export default function GeneratedGuiContainer({
   containerId: string;
   viewer?: ViewerContextContents;
 }) {
-  if (viewer === undefined) {
-    viewer = React.useContext(ViewerContext)!;
-  }
-  if (viewer === null) {
-    return null;
-  }
+  if (viewer === undefined) viewer = React.useContext(ViewerContext)!;
+
   const guiIdSet = viewer.useGui(
-    (state) => state.guiIdSetFromContainerId[containerId],
+    (state) => state.guiIdSetFromContainerId[containerId]
   );
   const guiConfigFromId = viewer.useGui((state) => state.guiConfigFromId);
 
@@ -427,7 +423,7 @@ function VectorInput(
         precision: number;
         onChange: (value: number[]) => void;
         disabled: boolean;
-      },
+      }
 ) {
   return (
     <Flex justify="space-between" style={{ columnGap: "0.3rem" }}>
